+++ changelog.md
@@ -30,6 +30,7 @@
 * **[Doctrine]** Ability to customize the startup time allocated to the player
 * **[Mission Generation]** Ability to choose whether player flights can spawn on the sixpack or not
 * **[Options]** New options in Mission Generator section: Limit AI radio callouts & Suppress AI radio callouts.
+* **[Options]** New option to use the combat landing flag in the landing waypoint task for helicopters.
 
 ## Fixes
 * **[UI/UX]** A-10A flights can be edited again